I added a spec file named ...no_eos2.
Building the RPMs for the hdfeos2 version is a bit of a pain, but
I made it a bit simpler by adding a special spec file for the straight
build and hacking the hdfeos build so it makes rpms with hdfeos2
in their names. It looks for the hdfeos2 library in $prefix/deps.