Fixed test issues in dapreader
These issues were actually due to a subtle change in the DAP4 function
parser. Sting constants are now returned with their double quotes.
This is so that variable names can be passed into functions even when
those names are quoted (because they contain spaces or meta
characters). The tests assumed unquoted names. I wonder about this,
but I'm checking in the fix and adding a note to jira/trac.