The dapreader 'error' tests use a pattern rule now.
A persistent problem with the tests here that check the parser's
detection of errors was that the baselines were compared to
the output using a litteral match. When the source line numbers
changed, the tests failed (i.e., every time the code was
edited). I switched to an AT_...PATTERN... test driver, modified
the baselines to contain just the txt of the error message,and
the problem went away.