Fixed dapreader so that it supports multiple containers.
When reading from a dap2 or dap4 data response, the code in
Connect is not 'container safe' so this handler reads into a
temporary DDS and then transfers the resulting variables. At
this point, processing global attributes is not handled for
a multi-container response (but it's hard to know how that
should be handled by this code).