Simple tests of roi() using installed handlers.
This is the first of the end-to-end tests for the server
function. These test if the function runs using the
dap module evaluator (different than using besstandalone
or a unit-test!) and if the output can be chained
to other transmitters like fileout_netcdf (it can).
The tests work, but I have not 'made them automatic'
yet because I'm not sure what that would mean for
typical builds (i.e., I think the the BES has to be
installed for these to work).