Added Odometer class; modified tests.
Odometer will be used to make a general bbox() function - one that works
for arrays with any number of dimensions and ne that can be used to
make minimal boxes that are the unions of regions, not other boxes.
I also tweaked the test code and marked most of the regression tests
as xfail since they will until we fix dapreader (Trac #2286). Lastly,
I added a subset version of the Modis 04 file that's easier to use
for tests since it's possible to figure out if the bbox/roi code is
working.