Copied the roi() tests to standalone (from cmdln)
I'm starting to think that we should have a special dir for
end-to-end tests where the whole server is used but still does
not have to be installed. By making the dir separate from the
code, we can make sure it is listed last and/or not listed
if the modules are not built as part of the bes build.