Fixes for building on CentOS 6
Most of the build issues cover #include filenames where
OSX is not case sensitive but linux is. Some other issues
are somewhat perplexing (a missing call to generate package.m4).
Also, the tests for the dap module contain some baselines
that use hash values that are machine/os specific and I've
removed those for now.