Added support for different dimensions to tabular()
The first version of tabular() required all arrays to be the same size.
The new version supports the case where one or more array is one 
dimension larger than the rest of the arrays. The tests pass, but there
is still considerable work to do for the code to be ready for use.