Modifications to make a 'NASA' rpm
This will include hdfeos2 and hdf4 statically linked to the
hdf4_handler because that's what has to be. I have added a new
target in the Makefile (nasarpm) that will do this and made
changes to configure so that a partial build of the dependencies
will provide the stuff needed.

Note that the hdf4 library is still not linking statically! Return
to this ASAP.