Makefile.am now supports a new 'all-static' rpm.
This target (all-static-rpm) expects to find the dependencies for
all of the handlers either on a plain RHEL6 machine or in the special
static dependencies installed in $prefix/deps where $prefix is set
in the shell and used as the value of --prefix to the hyrax-dependencies
project. This requires that the hyrax-dependencies be built with this
static linking in mind.