Fixed problems in besd and bescmdln
The besd script failed with installed from the rpm because the
bes prefix was set to /usr and then passed to besctl, which broke
all kinds of things. I removed that and the unused exec_prefix
and 'source functions' lines. Also, bescmdln logic made it so
that the value given with -p was ignored.