Modifications to MakeMaskFunction
The code compiles. To get that I used #if 0 ... endif to remove some of
the code (because it called functions that haven't been written). I
added two new functions that I hope will make it easier to build up
information needed to set pixels in the mask. I added unit tests for
those new functions.