Modified make_mask()
This version of make_mask() no longer takes a Grid size as its first
two arguments; now it takes a 'shape string' that provides the number
and size of each dimension of the resulting mask. However, I realized
that was probably not needed since that same info can come from the
dimensions that are passed in. I'm commiting this version, all the
same.