I patched the call to Split()
HDFCFUtil::Split() expected that the arrribute text (well text, it
doesn't know it's an attribute) is null terminated. However, the
text comes from a vector<char> and I can see the 16 calls in HDFSP.cc
that assign to Attribute::value are not appending a null - at least
the ones I looked at. My patch/hack copies the text, glues on a null
and then calls Split(). This works fine since the result of Split()
is a new variable. The valgrind off-by-one errors are now gone.