Added the 6.19.15 baselines
These files contain performance data from an early run of these
tests. There were a number of problems with the handlers/datasets
that had to be sorted out, but I thought it was important to save
this data. The raw logs are huge (many files > 500M) so I'll not
put those in github now (but they are backed up using an AMI
snapshot).