Compile, test and deploy BES using standard configuration
Attempting to build a standard configuration of the BES using just the
modules that I need and without all of the dependencies that I don't
need. So using the configure_standard.ac configuration file.

The besd.in file which builds the besd initialization script that goes
into /etc/init.d was not setting up the prefix variables correctly so
the bes was not launching using that script.

configure_standard.ac seemed to have not been updated with the other
configuration file as it appears to be working.

dapreader Makefile.am was missing the inclusion of the dap server libs
to resolve the Ancillary methods used, specificall reading a das
ancillary file.