New feature: Use BES.ServerIP to specify an IP on which to listen
This was requested by mmceniry@itsc.uah.edu (see https://opendap.atlassian.net/browse/HR-47).
If not given, the server reverts to its old behavior of listening
on the default address. Some of the files' edits are simply reformatting.
In those cases where most of the file was reformatted and I also
made a functional change, I put my initials and the date in a comment.