Hacked BESDapResponseBuilder.cc so functions returning structures work
A test hack: If we see that a server function has returned a Structure,
assume that it should be 'flattened' and do so. This means that the
ugrid() subsetting function can return its multiple variables and the
client will see that as such and not a Structure with N fields. This
preserves the look of the source dataset. We'll have to think about
this hack and how best to really implement it. The code could be faster
(using fewer/no copies) and could be less prone to failure (it assumes
every Structure returned by a function should be flattened). Not
insurmountable problems...