Adding timeout to BESInterface
This is to address Jira ticket BES-38 where the folks at
LARC (MMIC) want Hyrax to timeout when a function takes
too long. I'm using alarm() and trying to set it up so
that a 'context' can be used to pass a timeout value from
the OLFS to the BES (to minimize config file confusion).
It doesn't work yet...