Added error test bescmd files to bes/dap - not used, however
The autotest tests in bes/dap use dapreader, but that is not
built until after bes/dap is built. I moved two tests into
bes/dap/tests all the same because that's the best place to
not loose them. At some point we'll need to refactor the bes/dap
testing and get it working again and these will be useful.