BESDapResponseCache works!
This code uses a specialization of Sequence (CachedSequence) that
reads values, row by row, from the object's own internal data.
It uses the Sequence::serailize() and intern_data() implementations
so that nested sequences are sipported. The specialization of
Sequence might also serve as a replacement to the TabularSequence
code in bes/functions.