Fix for an errant test in H4ByteStream::read()
The code was testing if the correct number of bytes were read, but
then we added support for compressed chunks. Because the test was made
on the read buffer size _after_ the dat were decompressed, it failed
since the bytes in the buffer were larger than the bytes read off
disk. I moved the test so it's before the call to decompress the data,
and the test now passes. Unresolved is the issue of just what the
d_bytes_read field means - is it the # of bytes read from disk or the
number of bytes in the buffer. I have left it at the latter.