Fix to BESFSDir for the opendir/readdir race condition.
When two processes are using the same directory and one process makes
and deletes files in the directory at the same time as the other 
process uses opendir/readdir to iterate over the files in the
directory, it's possible for file names to show up in entries returned
by readdir that no longer exist (i.e., stat(2) returns -1). It's better
to ignore this than signal an error (IMHO).