Refactor in the gdal_handler
Extract functions for building global and variable attributes
Make a new header for the 'gdal utils' and move the util code into 
a source file with than name.
Make GDAL library handles const & paramters
Remove old code.
Remove files used with this was a separate project - this will serve as
a test for the checklist in ticket Hyrax-401 (or parts of that ticket).