New version of gdal_build_dmr() is much better at making shared dims
Also, lots of new tests; all pass. This version of the handler will
not take care of reduced resolution bands as gracefully as could be
the case, but it's pretty good (it will not bind them to shared dims).
The DMR is correct and build from DMR/DAP4 API calls except for the
attributes which are made using DAP2 AttrTable objects that are transformed
to DAP4 attributes.