Massive changes - I re-worked the DAP_MODULES conditional build
DAP_MODULES --> DAP_BUILTIN_MODULES and is now olny used in the
main bes Makefile.am. The builtin modules only work as such. The
DAP_EXP_MODULES symbol/conditional is reserved for the future.
I modified the hdf4/5 modules minimally, but had to update the
symbol name in their code.