Found the likely problem with hello_world and distcheck.
Symptom: distcheck seems to be putting generated files in
the srcdir and not builddir - this includes things like
bes.conf. The problem was (maybe - we'll see how far the
build gets) that hello_world (the directory) was included
in a parent Makefile.am's EXTRA_DIST and having generated
files included in the distribution means they are put in
$srcdir and not $builddir. And apparently including a dir
in EXTRA_DIST means put everything in the dist, even if
that dir has a Makefile.am.