Modified dap/unit-tests/test_config.h so that it includes the build dir.
This fixes the distcheck target. It works because the build dir
is where all generated files go and the bes.conf is a generated
file. The TemporaryFileTest.cc code needs to use that to reference
where bes.conf will be. With distcheck, it's in the special _build
dir; with a normal build, it's in the source dir.