Trying to sort out the ShowPathInfo test files that use sym links
Using sym links checked into git breaks distcheck because
tar copies the file/dir linked to, not the link itself. So
tests that should fail because the BES should not, as part
of the test, follow sym links will not fail because the 'link'
is not actaully a link anymore.