Fixed tests for get_site_map.
I realized that I needed to use different catalogs for the different
tests because the CatalogUtils objects are singletons and do not
'pick up' new valus for the keys. Thus, changing the value for a
catalog by editing the keys will show up in the catalog object, but
not the (singleton) Utils object. ...no idea why the Utils are singletons.