checkpoint. What is here, works. The tests pass.
The CLI tool build_dmrpp will print a DMR++ with chunk info, where the
chunk info is read using the HDF5 API. Inputs are an HDF5 file and
a (plain) DMR. The result is a DMR++. The handler has the capability
to print/return _either_ a DMR or DMR++ depending on the value of
the class static bool dmrpp::DmrppCommon::d_print_chunks, which
defaults to false. Yes, I know. A global. An alternative would be
to load up the Dmrpp<Type> objects with chunk info, and then insert
the corresponding elements into the DMR DOM tree. Version 2?