Added redirection to the DMR++ handler for data requests when that's appropriate
The DAP2 and DAP4 data requests look in the MDS for DMR++ responses
that match the relative path of the current container. If one is
found, the container type is set to dmrpp and container attributes
are set so that the DMR++ handler will realize that it should look
in the MDS for a DMR++ response (normally it uses the value of access()
to find a DMR++ file on disk).

Still to do: implement the DAP2 code in the DMR++ handler and tests.