Added tests for the DMR++ changes to the BES framework and DMR++ handler.
Only one file, but the changes to the framework, such that a file with
a DMR++ response in the MDS is redirected to the DMR++ handler, clearly
work. More tests won't prove much, but the change should use a bes.conf
key to disable the redirection since it will access the MDS and if an
admin knows the DMR++ is not being used, why do that for every request?