Resolved the missing xml:base issue.
If the BES contxt xml:base is missing do not throw an exception
or use a default value. Instead, do nothing WRT the xml:base
attribute. If the attribute was present when the response was
added to the store, that value wil be present when it is extracted.
If no value was present, there will be no value in the response
that is returned.

The OLFS should _never_ call the BES without setting the xml:base
context, so all responses from teh MDS should have this attribute.