BESCatlalogUtils hacking.
Fixed tests. One side effect of the modifications I've made is that
unit tests that call BESCatalogUtils() must have the keys for the
default catalog, its RootDir and TypeMatch set. This only affected
four tests, but it's not the best design. Hopefully this can be
fixed in a later refactor.