Fixed ticket HYRAX-802.
We were pretty close with our 'send_p' hack, but there was one wrinkle -
the send_p flag has to be cleared after the data are read in case a CE
is then applied. Failure to do so leaves the CE parser thinking it's
working on a variable that has already been constrained. That's OK,
but only in limited cases.
I also modified the tests/Makefile.am so that TESTSUITEFLAGS env var
works (it was overridden in the Makefile).