Fixes for HYRAX-803 - these changes patch a memory error flagged by ASAN.
The problem appears to be in the copy constructor(s) for the two classes
ArrayAggregateOnOuterDimension and ArrayAggreagtionBase or is a latent
problem they trigger (although those methods are so thin it's hard
to imagine them being broken). Regardless, replacing the DDS::add_var()
calls with DDS::add_var_nocopy() calls fixes the errors. See the bug
report for an extensive discussion of the errors.