Modified travis.yaml to install/start the bes before running the tests
This will run the cmdln tests, which test the server and not the
besstandalone app, which is just a tiny bit different. In this set of
tests the server is installed and will load all of the modules in
their default configuration. The besstandalone tool only ever loads a
subset of the modules and often with 'special' settings.