The MDS does not work with server functions.
When a CE contains a server function call, the MDS should not be used.
In that case, the MDS will try to build a DDS and evaluate the CE, but
the type classes in the DDS won't support data read operations, resulting
in an obscure error about unimplemented read() methods. This commit
fixes the problem for DDS and DMR responses, but not DDX responses.
The latter don't appear to handle constrained requests at all.