Changed logging of errors
Added BES.ExitOnInternalError option, If true, the BES will
exit on an InternalError (default is the current behavior,
which is to not exit).

Added BES.DoNotLogSourceFilenames which defaults to false.
Now BESError's will be logged with the filename and line number
by default (new behavior).

Also 'improved' instrumentation in BESFileLockingCache making it
a bit easier to see what's going on. Still needs work.