Annotation service: Added tests along with a fix for the 'container collision' bug
Handlers do not generally test to see if an attribute container exists
before making one, so if the Annotation URL attribute is added to
DODS_EXTRA and then the handlers are run and one of those handlers
also uses a container called DODS_EXTRA... libdap throws. My fix is
to call the handlers first, and then add the attribute. I look for
DODS_EXTRA and insert the attribute or make the container, as needed.