Fixed HK-313
If a Remote Resource was accessed using only the collection name with no
trailing slash, an exception was thrown because the string was accessed
one character past its end. Now it's OK to leave off the trailing slash.
I also found that the BESUtil::split() function is broken and will fix
it in a subsequent commit.