Partial fix for the contiguous chunk parallel read code.
Two test still fail. The error was that we used the whole value for the
child chunk offset when determining where to put a child chunk's data
into the master chunk. This resulted in writing data far past the end
of the master chunk's buffer. The correction is to subtract the master
chunk offset from the child chunk offset. At least, that fixes most of
the tests.