Updated travis.yml: removed the www.opendap.org deploy operations
Not really sure why this stopped working, but maybe the www.o.o certs
changed because the build is failing when it can no longer verify the
www.o.o host. We might use an S3 event to trigger the copy of these to
the host, thus moving it out of travis' realm.