Added methods to TheBESKeys for getting and setting vectors of values associated with a key and for getting and setting a map associated with a key.
Added  code to ngap_module to HEAD the data access
URL retrieved from CMR and record the last accessed
URL (which may be a signed AWS s3 request URL)
in TheBESKeys

Addec code to the dmrpp_module Chunk class to retrieve
the target URL (last redirect accessed url) from the cache
and us it to make requests.

Adly, this is broken atm.