I moved the package operation ahead of the scan
This way we can find out if there's a CentOS issue before the long 30
minutes of scanning. The code is not deployed for a PR and not if the
scan fails, so we still meet the criteria that a failed scan prevents
deployment (or merge to master, given our basic rule that a PR must
pass before it can be merged).
[skip ci]