Csv handler fix (#462)
Looking for the mysterious 'syntax error' messages.
Added tests. There's plenty of room for improvement here...

The file name (actually a pathanme)  has to be removed from the 
baseline since it is different for check and distcheck. The regex used 
when forming the baseline was not including underscores. Now it does.