Fixed the obscure failure of tests 27, 28, 78 and 79
What's going on: When this code finds Grids, it knows that the
Maps of a DAP2 Grid object/type should be represented in a netCDF
file as a 'dimension' and that dimension might be shared by several
netCDF arrays. For those 'special case' DAP2 Arrays (those ones that
are Maps in a Grid) we cannot delete their storage until the whole
dataset has been written to the netCDF file. It's a drag, but Maps
are typically small. DAP4 Shared Dimensions should not have this
problem.