Fonc refactor v2 (#505)
Initial step at a streaming response.

This branch holds code that streams data to the built netCDF3/4 file one variable
at a time. The memory used is never greater than the largest variable plus the size
of any DAP2 Grid Map (which have to be held in memory for the duration of the 
serialization process). In addition, we have attempted to eliminate all unneeded 
copying of data values, writing values to the netCDF3/4 file directly from the DAP
variables.

Note that for the NC_INT and NC_SHORT code: When it can use the

Co-authored-by: The-Robot-Travis <npotter@opendap.org>
Co-authored-by: James Gallagher <jgallagher@opendap.org>
Co-authored-by: Kent Yang <myang6@hdfgroup.org>