Added a bit to the fletcher32 checksum computation.
I found that the checksum might not be computed correctly since the
value we read does not seem to match the computed value. We can pick
this up next week. The data values read for the test _do_ match the
vaues in the data set. Could this be that the checksum needs to have
its word order twiddled depending on the machine's endianness?