Fletcher32 (#536)
* first implementation of fletcher32 support

* first implementation of fletcher32 support

* updated a comment in DmrppCommon.cc

* second implementation of fletcher32 support

* adding parsing of fletcher32 filter

* Added tempated f32 checksum stuff

* adding parsing of fletcher32 filter

* adding parsing of fletcher32 filter

* added parsing of fletcher32 filter

* added fletcher32 test

* added fletcher32 test #2

* Added a bit to the fletcher32 checksum computation.

I found that the checksum might not be computed correctly since the
value we read does not seem to match the computed value. We can pick
this up next week. The data values read for the test _do_ match the
vaues in the data set. Could this be that the checksum needs to have
its word order twiddled depending on the machine's endianness?

* Changed the fletcher32 algorithm code to match the HDF5 version

Also added source pragmas to suppress warnings about alignment and
asserts to test that the data are correctly aligned at least when the
tests are run.

* added working fletcher32 checksum

Co-authored-by: The-Robot-Travis <npotter@opendap.org>