load_attributes() working for the test cases
Tests include simple arrays, arrays of structres and
datasets with nested groups. This implementation of
load_attributes() searches for the location in the DOM
tree using a chain of variables built from the parent
back pointers. A different (better?) way would be to
build a cache of variables to xml_nodes, but that requires
that the 'thin' DMR be built first by this code. Maybe
we (can/will) get the DMT some other way...?