Changed how the code links to the dap_module to fix 'non-portable' warnings (#569)
* Changed how the code links to the dap_module to fix 'non-portable' warnings.

The code was linking against dap_module.la which becomes dap_module.so
which is actually a 'module' and not a real library. This generates
warnings from libtool and also, on linux, with rpath. The changes are
mostly in unit-tests but also appears in the dmrpp tools like build_dmrpp.

* Fix for centos 7 rpm and static archives

The RPM build used --disable-static, but the code now links with
libdap_module.a. I removed --disable-static. However, the .a archives
are not part of the RPM, so the install step removes them.