Turned off the 'stop escaping attributes' modification
It turns out that this is not needed for the history_json attribute.
That was hard to see, but haiving this on have _no effect_ on the
operation of that code.

I also reorganized the tests so there is one testsuite, not two since
the time when netcdf4 support can be optional has long passed.