Non-fix for the 'MDS Bug' which was actually an error in the netcdf handler
A change was added there to accommodate the 'history_json' attribute
but then (I think) the design changed and the code to stop escaping
attributes was not needed (which is good, because it breaks stuff like
the MDS). The history_json attribute was implemented entirely in the
fileout_netcdf handler and so the change in the netcdf_handler was moot.