History mds fix (#586)
Turned off the 'stop escaping attributes' modification:

It turns out that this is not needed for the  history_json attribute. That was hard to see, but having 
this on have _no effect_ on the operation of that code. See HYRAX-649 for more info

I also reorganized the tests so there is one testsuite, not two since the time when netcdf4 support
can be optional has long passed.

Removed conditional support for netcdf4 - it is now required

Removed the old netcdf_handlerTest.at in favor of testsuite.at