Dmrpp string array fix (#587)
Patch for HYRAX-451 - String Array failure.

This addresses only the simplest case where we have a String Array
with one element and no constraint. Other cases will fail. This
case needs tests. Then we will need to expand the way String
Arrays are represented in the DMR++ and modify the handler to process
the various ways string arrays can be encoded.