Added a new BES key to the hdf5 handler EscapeUTF8Attr (default true)
When true, use the same escaping scheme for UTF-8 values for attributes
as for ASCII. If false, allow the UTF-8 values to appear unescaped in
DAS, DDX and DMR responses. Those values currently cause failures in the
OLFS when it builds the HTML form from the DMR or DDX (the failure actually
probably happens in the browser and me be due to the xml character encoding
we use for the DDX and DMR XML which is ISO-8859-1).