Fixed hdf5_handlerTests build in the Makefile.am
The autotest shell script is built using autoconf and thus is included
in the distribution tar ball (because users should not need to have autoconf
to build the code). The bes conf files, however, are built when the user
compiles. Thus, those conf files cannot be depencies of the testsuite
script but instead should be dependencies of the check target.