Esc attr 536 delta (#593)
Added a new BES key to the hdf5 handler EscapeUTF8Attr (default true)

When true, use the same escaping scheme for UTF-8 values for attributes
as for ASCII. If false, allow the UTF-8 values to appear unescaped in
DAS, DDX and DMR responses. Those values currently cause failures in the
OLFS when it builds the HTML form from the DMR or DDX (the failure actually
probably happens in the browser and me be due to the xml character encoding
we use for the DDX and DMR XML which is ISO-8859-1).

The tests include a new macro that I'd like to also add to the common
collection of macros in bes/modules/handler_tests_macros.m4. It provides
a way to test different bes.conf files against the same bescmd file
without using multiple .at files. As a bonus, this also shows how to
'inherit' from a bes.conf file to avoid repeating listings of keys and
how to avoid the need to build a conf file using configure/Malefile.am.