Error redux (#598)
Edits to the BESError and child classes - inherits from std::exception now

Note: gcc (i.e., linux, for us) does not like the default copy constructor (but OSX/cland does).
Having the default copy constructor 'violates' the rule of zero and Sonar was griping about
that. I marked it as a false positive.