Auto ptr begone (#600)
Removing auto_ptr from all but the NCML handler

In the course of doing this, I modified the sonarscan properties file in what seemed
to be an innocuous way yet the scan broke with a message about node.js. I removed the
change and the error went away. The change was to remove a directory that no longer
exists from the 'exclude' list. No Idea why this broke the scan, but it was the only change 
and undoing that 'fixed' the problem.