Size fix (#604)
Extended the response size code so that Apple/OSX values are correct

I also moved the code into a new DapUtils file that can be
used by Transmitters outside of the bes/dap code. This way
the ASCII, etc., code can log the size info. I also made
DapUtils a namespace and not a class with static methods.