Makefile.am was 'cleaning' the testsuite scripts
Cleaning these scripts with a parallel make meant that
in some cases the testsuite script was removed before it
could be run as per 'clean-local' resulting a failure.
Since the testsuite script is included in the tar ball,
I removed it from CLEANFILES in three Makefile.am files.
This seems to fix the issues with 'make clean --jobs=12'
failing.