Hdf5 fill value (#616)
Add support for HDF5 fill values to the DMR++ XML files and to the DMR++ Handler

When fill values were used by HDF5 to eliminate storing 'empty' chunks, the 
DMR++ handler was choking since it was trying to read data from chunks that
did not exist. 

This change has two parts: build_dmrpp/get_dmrpp now builds DMR++ XML 
files that include information about the fill values, and the DMR++ handler will 
use that information to build the correct data values for those chunks which 
use the fill values. 

There are some cases that are not currently supported such as string variables, 
hdf5 array variables (which are different from 'datasets' (that's an array in DAP) 
and compound variables.