New gdal (#646)
* Added modules/configured_featres.txt

This can be used in AT_SKIP_IF() to make our tests sensitive
to features.

* Added a 'skip this test' based on a features list to the handler test macros

And used that for the function module when the server si built w/o
gdal.

* Modified ncml handler to skip tests if a feature is not present

More of a hack, really. This only works for the macros AT_CHECK_ALL_DAP_RESPONSES
and AT_CHECK_DATADDS_GETDAP but that is good enough for the tests that
use gdal. I think we should probably change AT_CHECK_ALL_DAP_RESPONSES_WITH_CONSTRAINT
but I don't know how to test that w/o much more work. Suggest a new ticket.

* Fixed the tests adnd modified AT_CHECK_ALL_DAP_RESPONSES_WITH_CONSTRAINT

The AT_CHECK_ALL_DAP_RESPONSES_WITH_CONSTRAINT macro now supports
skipping tests if a feature is not present.

* Added conditional bes.conf use to Makefile.am

There are two different bes.conf files depending on whether gdal
is present. This is probably not scaleable to O(10) features but
for one or two it should work OK.

* Added an experimental gdal_module that encompasses both read and writer

This will enable the functionality now spread across two modules
in one, so that we only have to link once to gdal and reducing the
module count in the process.

* WIP

* WIP

* The reader works

* Both the reader and writer functions work now.

* Clean up after the hackery

* Removed GDAL from functions for now

* Removed CPLQuietErrorHandler from the functions module

* moved gdal_handler and fileout_gdal to modules/unused

* distcheck WIP

* distcheck WIP

* more distcheck hackery

* Back out some parts of the 'make gdal optional' features from ncml_module

* Removed the if BUILD_GDAL ... endif lines from Makefile.am files

* More cleaning in functions

* Fixed an error WRT GDAL and ScaleGrid.h

* Odd linking issue. Fix trial

* Revert functions/unit-tests/Makefile.am to the version on master

* Remove GDAL-dependent unit tests from the functions module

* Moved STARE template functions to StareFunctions.h

* Warnings & cleanup in the STARE functions

* The cast for STARE_SpatialIntervals to dods_uint64* is needed on RHEL8

* more cleanup

* Trigger TravisCI

* Trigger TravisCI

* Revert Travis change [skip-ci]

Co-authored-by: ndp-opendap <ndp@opendap.org>