Remote Resources bug 2 (#701)
Fixes for a concurrency issue reported by NSIDC.

Removed the header caching and cached item expiration features from the
RemoteResource class. The headers were not being used, however, we will
need to address this issue in the future. The Expiration Issue needs to be addressed
before this code can go in production. We can test this patch in UAT to see
if the problem reported by NSIDC has been addressed.