Bes keys fix (#733)
TheBESKeys --> unique_ptr pattern

* TheBESKeys now uses a unique_ptr and passes tests.

Some other code had to be changed because the singleton
was 'messed with' by some of the BES framework destructors
and some of the unit-tests. I added a reload_keys() method
to enable unit-tests to switch key files and/or reload the
original key file, effectively resetting changes they might
have made during the tests.