Frankenbranch (#752)
The Build DMR++ module and the RemoteResource changes. 

This merge includes the additions for the Build DMR++ module of the BES
and the changes to the bes/http convenience library that removes caching
from the RemoteResource class. Other changes include the way RR is used
in various modules in the BES, including the NGAP, Gateway, and HTTP
catalog modules.