I modified the get_dmrpp CLI tests to include two tests for -r
The -r option is the 'reduce' option where a single file for
missing values can be used for more than one data file. One of two
of the new tests work. I also added a very verbose option so that the
versbose option by itself is easier to use while retaining the
more complete output of the orignal version.