RemoteResource retry fix (#771)
* Fixed the RemoteResources retry code

The problem was that ftruncate was used to remove an error
message from the file used to hold the response as it was
read from the network. However, the file descriptor was not
rewound to the start of the file, so the valid response data
was written starting past the initial position. A reader would see the
initial part of the temporary file as filled with nulls. This
error was introduced with RR was rewritten.

Also refactored the CmrApiTest.cc code since this work overlapped
the CMR legacy API and I was working on those tests.

Added tests for the RR retry fix