S3 module (#773)
An 's3 reader' module for the BES. To be used for testing...

This could also be the basis for a module that provides access to data 
held in S3 in a way that is more general than the current module that relies
on CMR for the locations of the data files.

This branch also contains fixes for issues in the 'http' convenience library:
 * Refactor of CredentialsManager so that load_credentials() is always called
 * Pass the vector<char> buffer in to the caller so that memory does not go 
out of scope once the function returns. I suspect that rapid_json (and similar
libraries) depend on the document memory.
 * We remove the include of rapid json in CurlUtils.h The CurlUtils.h header
is installed but the rapidjson headers are not. Remove the http_get_as_json() 
function and use the simple http_get() and have the client build JSON however 
they want/can.
* Removed NgapCredentials since it's not used.

---------

Co-authored-by: Samuel Lloyd <slloyd@opendap.org>