Changes - fixed leak (unverified as of now) (#788)
Fixed a leak in the build DMR++ module unload

Also refactored to adopt C++14 features.
Added a test for the DMR++ builder that will work in more places.
It still requires access to ESDIS CMR and a granule name that exists
in CMR, but it is not local to one machine. The test is not run by 
default, but is included and can be run by hand if CMR access is available.