Flc thread test (#792)
Removed the lock_cache_write/read() and replaced it with a RAII class

I removed the exceptions that were thrown as messages. Now error 
messages are logged.Dropped the fancy error message in the new 
AdvisoryLockGuard class' constructor and destructor since there are 
OSX -vs- Linux issues with getting a file name from a descriptor.