Changed DMR++ object memory cache so it is off by default
The cache seems to have an error in the D4Maps code. Also fixed an
error in the test code that mistakenly used a baseline file as if it
was the file to test. And added missing code to read the value of the
keys used to configure the cache - so two errors in the code were
fixed in addition to changing the default behavior.