Added DMR++ 'content' caching to the NgapContainer
I used the existing BESContainer attributes to signal that access()
was going to return the cached content (the whole XML doc as a string)
and not a file name. I extended teh DMZ parser so that it can parse
strings. I added some timing to the new code along with 'hit' or 'miss'
text for the 'cache' BESDebug key. It's pretty rough still.