This code has the home grown CurlHandlePool removed and the libcurl 'share' using in its place
For this modification, the code that managed the curl handle pool (the original
code) processed data from 20 requests for 2 variables with contiguous storage in
222ms. Removing the 'handle reuse' of the pool saw that time change to 351ms.
Using the sharing feature of curl (and the blunt force trauma lock functions)
and that time becomes 227ms. So our self-managed and the libcurl scheme are
effectively equal, with the latter having some room for better performance if
the lock functions are improved. jhrg 10/6/23