Added get() and tested get/put collisions are handled
In this case, 'handled' means that the cached item is not broken.
If a call to get and put collide, then if get goes first it fails
and then put() runs and then a second call to get() works. If put()
goes first, it blocks get() and completes and then get() works.