File cache (#837)
This is a simple cache for files. Its API is limited to get(key, &item), put(key, filename),
del(key), clear(), purge(), and initialize(...). For all of these operations, the cache is locked.
The get() operation returns _item_ with a valid open file descriptor to the file that is
locked using a shared lock. When item is destroyed, the lock is cleared. There is a second
version of put() that takes/returns a &PutItem that holds an open file descriptor that other
code can write to. When the PutItem instance is destroyed, the fd is closed and the exclusive
lock on the file is released. The cache is process- and thread-safe.