Ngap file cache (#847)
This is the culmination of the work on DMR++ caching.

A memory cache is used by each process to hold DMR++ documents for fast reuse.
Timing data shows the memory cache returns hits in ~100uS. This is backed up by a
file cache that can be accessed across all the processes on a given host (and across
hosts if the file system supports NFS - which AWS's EFS does). Timing data shows
cache hits for the file cache return items in ~1mS. Network access has been refactored
so that the DMR++ is read from the wire directly into a C++ string object and this takes
about 500mS without the overhead of authentication (i.e., in tests where we use OPeNDAP's
S3 bucket and do not rely on a web service for credentials or 'URL signing').