These changes patch the issue in HYRAX-1334
The problem is that over time the software has moved from the
original FONcTransform constructors that initialized the FONc types'
DDS or DMR fields (and the constraint expression evaluator field)
and so those can be null. The intern_data() method was being called
using references to 'nul objects' which oddly didn't seem to
hurt the way the code ran, but was not good. This patch works around
that. A better fix would be to refactor the underlying code and
fix the logic that uses objects that are apparently not needed.