Fonc asan patch (#881)
These changes patch the issue in HYRAX-1334

The problem is that over time the software has moved from the
original FONcTransform constructors, which initialized the FONc types'
DDS, DMR, ...,  fields so that those can now be be null. Because of this
change, the intern_data()  method was being called using references to 
'null objects' which oddly didn't seem to hurt the way the code ran, 
but was not good. 

A better fix would be to refactor the underlying code and fix the logic.